Add LoginForm component
Form functionality is here separated from the Panel. We load what data we need from app config and then when the button is pressed we log the user in with Keycloak. If they are already logged in, this will take them straight through to COP, either Staging or Prod as per current app config. If not, it will take them to the POISE Keycloak login page, bypassing our theme page.

idpHint environment variable is added to identify which identity provider we would like to use—in our case POISE.